+++ WordbookImpressApp/WordbookImpressLibrary/Storage/WordbooksImpressInfoStorage.cs
@@ -23,6 +23,14 @@ namespace WordbookImpressLibrary.Storage
             Content.Add(item);
         }
 
+        public static void Init()
+        {
+            if (System.IO.File.Exists(PathBup)) { System.IO.File.Delete(PathBup); }
+            if (System.IO.File.Exists(Path)) { System.IO.File.Delete(Path); }
+            Content = new ObservableCollection<WordbookImpressInfo>();
+            OnUpdated();
+        }
+
         public static async Task<WordbookImpressInfo[]> LoadLocalData()
         {
             if (!System.IO.File.Exists(Path))
@@ -34,19 +42,26 @@ namespace WordbookImpressLibrary.Storage
             try
             {
                 result = await Helper.SerializationHelper.DeserializeAsync<WordbookImpressInfo[]>(Path);
+                if (result != null)
+                {
+                    Content = new ObservableCollection<WordbookImpressInfo>(result);
+                    OnUpdated();
+                    return result;
+                }
             }
             catch
             {
-                try
-                {
-                    result = await Helper.SerializationHelper.DeserializeAsync<WordbookImpressInfo[]>(PathBup);
-                    if (System.IO.File.Exists(Path)) { System.IO.File.Delete(Path); }
-                    if (System.IO.File.Exists(PathBup)) { System.IO.File.Move(PathBup, Path); }
-                }
-                catch
-                {
-                    result = new WordbookImpressInfo[0];
-                }
+            }
+            try
+            {
+                result = await Helper.SerializationHelper.DeserializeAsync<WordbookImpressInfo[]>(PathBup);
+                if (result == null) return new WordbookImpressInfo[0];
+                if (System.IO.File.Exists(Path)) { System.IO.File.Delete(Path); }
+                if (System.IO.File.Exists(PathBup)) { System.IO.File.Move(PathBup, Path); }
+            }
+            catch
+            {
+                return new WordbookImpressInfo[0];
             }
 
             Content = new ObservableCollection<WordbookImpressInfo>(result);
